Factor out condor.record.util into condor.util
