+++ test/test_utils.py
@@ -1,6 +1,6 @@
 import pytest
 
-from condor.record.util import isi_text_to_dic
+from condor.util import isi_text_to_dic
 
 
 @pytest.fixture(scope='module')
